Lot of noise in this PR, sorry about that. What matters is that
* I'm using a 256byte table for complement calculation instead of map
* I'm avoiding heap allocations where possible by using a byte slice and unsafe conversion.

```go
// ReverseComplement takes the reverse complement of a sequence.
// ReverseComplement expects ASCII input.
func ReverseComplement(sequence string) string {
    n := len(sequence)
    newSeq := make([]byte, n)
    for i := 0; i < n; i++ {
        newSeq[i] = complementTable[sequence[n-i-1]]
    }
    // This is how strings.Builder works with the String() method. If Mr. Go says it's safe...
    return *(*string)(unsafe.Pointer(&newSeq))
}
```