@soypat 2022-11-27
Test is failing because I included a test for the complement table and it seems it was doing something I did not expect. 'U's complement is  'A'?