@Koeng101 2022-11-28
A few things to change (single letter variables) - mostly interested in benchmarks. How much faster is it?