@TimothyStiles 2022-11-28
I can fix variables and stuff but I'm also interested in benchmarks! Would be happy to start on a real benchmark suite for Poly @soypat.