@soypat 2022-11-28
@Koeng101  @TimothyStiles I'll fix the variable names and add better documentation today- but I'd like to note that `i` as a generic counter in small scopes has been shown to be as readable or even more readable than a full length name. See [Beniamini et al.](https://www.cs.huji.ac.il/w~feit/papers/SingleLetter17ICPC.pdf).

It'd be awesome to have a benchmark suite like you say to get a real taste of how performance changes affects those using the library :smile:, more than happy to help out however I can. Been looking at refactoring the `fasta` package :eyes:.

<details><summary>Benchmarks for 32byte string input (click)</summary>

#### Before
```
goos: linux
goarch: amd64
pkg: github.com/TimothyStiles/poly/transform
cpu: Intel(R) Core(TM) i5-8265U CPU @ 1.60GHz
BenchmarkReverseComplement-8   	 1639549	       714.7 ns/op	     224 B/op	       3 allocs/op
BenchmarkComplement-8          	 3199111	       383.4 ns/op	      48 B/op	       1 allocs/op
BenchmarkReverse-8             	 4345659	       325.0 ns/op	     176 B/op	       2 allocs/op
PASS
```
#### After
```
BenchmarkReverseComplement-8   	18890924	        60.14 ns/op	      32 B/op	       1 allocs/op
BenchmarkComplement-8          	25396846	        55.45 ns/op	      32 B/op	       1 allocs/op
BenchmarkReverse-8             	19949305	        55.67 ns/op	      32 B/op	       1 allocs/op
```

</details>